+++ src/stapi_fastapi/routers/root_router.py
@@ -7,7 +7,7 @@ from fastapi.datastructures import URL
 from returns.maybe import Maybe, Some
 from returns.result import Failure, Success
 
-from stapi_fastapi.backends.root_backend import RootBackend
+from stapi_fastapi.backends.root_backend import GetOrder, GetOrders, GetOrderStatuses
 from stapi_fastapi.constants import TYPE_GEOJSON, TYPE_JSON
 from stapi_fastapi.exceptions import NotFoundException
 from stapi_fastapi.models.conformance import CORE, Conformance
@@ -28,7 +28,9 @@ logger = logging.getLogger(__name__)
 class RootRouter(APIRouter):
     def __init__(
         self,
-        backend: RootBackend,
+        get_orders: GetOrders,
+        get_order: GetOrder,
+        get_order_statuses: GetOrderStatuses,
         conformances: list[str] = [CORE],
         name: str = "root",
         openapi_endpoint_name: str = "openapi",
@@ -37,7 +39,9 @@ class RootRouter(APIRouter):
         **kwargs,
     ) -> None:
         super().__init__(*args, **kwargs)
-        self.backend = backend
+        self._get_orders = get_orders
+        self._get_order = get_order
+        self._get_order_statuses = get_order_statuses
         self.name = name
         self.conformances = conformances
         self.openapi_endpoint_name = openapi_endpoint_name
@@ -164,7 +168,7 @@ class RootRouter(APIRouter):
             ),
         ]
         if end > 0 and end < len(self.product_ids):
-            links.append(self.pagination_link(request, self.product_ids[end]))
+            links.append(self.pagination_link(request, self.product_ids[end], limit))
         return ProductsCollection(
             products=[
                 self.product_routers[product_id].get_product(request)
@@ -177,14 +181,15 @@ class RootRouter(APIRouter):
         self, request: Request, next: str | None = None, limit: int = 10
     ) -> OrderCollection:
         links: list[Link] = []
-        match await self.backend.get_orders(request, next, limit):
-            case Success((orders, Some(pagination_token))):
+        match await self._get_orders(next, limit, request):
+            case Success((orders, maybe_pagination_token)):
                 for order in orders:
-                    order.links.append(self.order_link(request, order))
-                links.append(self.pagination_link(request, pagination_token))
-            case Success((orders, Nothing)):  # noqa: F841
-                for order in orders:
-                    order.links.append(self.order_link(request, order))
+                    order.links.extend(self.order_links(order, request))
+                match maybe_pagination_token:
+                    case Some(x):
+                        links.append(self.pagination_link(request, x, limit))
+                    case Maybe.empty:
+                        pass
             case Failure(ValueError()):
                 raise NotFoundException(detail="Error finding pagination token")
             case Failure(e):
@@ -204,9 +209,9 @@ class RootRouter(APIRouter):
         """
         Get details for order with `order_id`.
         """
-        match await self.backend.get_order(order_id, request):
+        match await self._get_order(order_id, request):
             case Success(Some(order)):
-                self.add_order_links(order, request)
+                order.links.extend(self.order_links(order, request))
                 return order
             case Success(Maybe.empty):
                 raise NotFoundException("Order not found")
@@ -231,10 +236,10 @@ class RootRouter(APIRouter):
         limit: int = 10,
     ) -> OrderStatuses:
         links: list[Link] = []
-        match await self.backend.get_order_statuses(order_id, request, next, limit):
+        match await self._get_order_statuses(order_id, next, limit, request):
             case Success((statuses, Some(pagination_token))):
                 links.append(self.order_statuses_link(request, order_id))
-                links.append(self.pagination_link(request, pagination_token))
+                links.append(self.pagination_link(request, pagination_token, limit))
             case Success((statuses, Nothing)):  # noqa: F841
                 links.append(self.order_statuses_link(request, order_id))
             case Failure(KeyError()):
@@ -267,28 +272,19 @@ class RootRouter(APIRouter):
     ) -> URL:
         return request.url_for(f"{self.name}:list-order-statuses", order_id=order_id)
 
-    def add_order_links(self, order: Order, request: Request):
-        order.links.append(
+    def order_links(self, order: Order, request: Request) -> list[Link]:
+        return [
             Link(
                 href=str(self.generate_order_href(request, order.id)),
                 rel="self",
                 type=TYPE_GEOJSON,
-            )
-        )
-        order.links.append(
+            ),
             Link(
                 href=str(self.generate_order_statuses_href(request, order.id)),
                 rel="monitor",
                 type=TYPE_JSON,
             ),
-        )
-
-    def order_link(self, request: Request, order: Order):
-        return Link(
-            href=str(request.url_for(f"{self.name}:get-order", order_id=order.id)),
-            rel="self",
-            type=TYPE_JSON,
-        )
+        ]
 
     def order_statuses_link(self, request: Request, order_id: str):
         return Link(
@@ -302,9 +298,11 @@ class RootRouter(APIRouter):
             type=TYPE_JSON,
         )
 
-    def pagination_link(self, request: Request, pagination_token: str):
+    def pagination_link(self, request: Request, pagination_token: str, limit: int):
         return Link(
-            href=str(request.url.include_query_params(next=pagination_token)),
+            href=str(
+                request.url.include_query_params(next=pagination_token, limit=limit)
+            ),
             rel="next",
             type=TYPE_JSON,
         )